Emphasize using verbs or participles to replace whitelist/blacklist
Languages other than English may not allow turning a noun into a verb with
the same agility that English has.  Some of them might not form compound
words easily and resort to English loan words.  In such a scenario,
it would not be impossible for the translator to revert "allowlist"
into a loan word based on "blacklist", because that's a more commonly
known word to his audience of non-English speakers.  This would make
the effort to use inclusive language vain.

Therefore, emphasize more the possibility to use not just the "-ed"
form of verbs, but also the verbs themselves without the "list" suffix.
Their past participles easily work as adjectives, and they are easily
translated.