Remove mention of origin of whitelist/blacklist in the publishing industry
This origin is not substantiated by sources such as the Oxford English
Dictionary.  The more likely origin seem to be in politics, where it was
applied to people, with the term's surge in popularity occurring during
World War I.  At the time, "blacklist" was used for the list of firms
denied access to various Entente infrastructure in the United Kingdom.
During and after World War II, the term started being used to describe
various kinds of restrictions, generally not limited to the publishing
industry.

See also:
https://books.google.com/ngrams/graph?content=blacklist&year_start=1800&year_end=2000

I tried keeping the reference to slavery, but I couldn't fit it in the
rest of the text which talked about the different metaphors behind colors
in various cultures.